+++ packages/language-c/grammars/tree-sitter-cpp/highlights.scm
@@ -84,6 +84,8 @@
 
 (primitive_type) @support.type.builtin.cpp
 
+(placeholder_type_specifier (auto) @support.type.builtin.auto.cpp)
+
 ; Mark all function definition types with data…
 (function_definition
   type: (_) @_IGNORE_
@@ -477,6 +479,8 @@
 
 
 ["&&" "||"] @keyword.operator.logical.cpp
+(binary_expression
+  ["and" "or"] @keyword.operator.logical.cpp)
 
 "++" @keyword.operator.increment.cpp
 "--" @keyword.operator.decrement.cpp
@@ -493,6 +497,7 @@
 
 (unary_expression
   ["+" "-" "!"] @keyword.operator.unary.cpp)
+(unary_expression "not" @keyword.operator.unary.cpp)
 
 (conditional_expression
   ["?" ":"] @keyword.operator.ternary.cpp)