+++ packages/language-javascript/lib/main.js
@@ -21,7 +21,7 @@ exports.activate = function () {
     content(callExpression) {
       const { lastChild } = callExpression;
       if (lastChild.type === 'template_string') {
-        return lastChild;
+        return stringFragmentsOfTemplateString(lastChild);
       }
     }
   });
@@ -29,8 +29,8 @@ exports.activate = function () {
   atom.grammars.addInjectionPoint('source.js', {
     type: 'assignment_expression',
 
-    language(callExpression) {
-      const { firstChild } = callExpression;
+    language(expression) {
+      const { firstChild } = expression;
       if (firstChild.type === 'member_expression') {
         if (firstChild.lastChild.text === 'innerHTML') {
           return 'html';
@@ -38,10 +38,10 @@ exports.activate = function () {
       }
     },
 
-    content(callExpression) {
-      const { lastChild } = callExpression;
+    content(expression) {
+      const { lastChild } = expression;
       if (lastChild.type === 'template_string') {
-        return lastChild;
+        return stringFragmentsOfTemplateString(lastChild);
       }
     }
   });
@@ -94,4 +94,10 @@ function languageStringForTemplateTag(tag) {
   } else {
     return tag;
   }
+}
+
+function stringFragmentsOfTemplateString(templateStringNode) {
+  return templateStringNode.children.filter(
+    c => c.type === 'string_fragment'
+  );
 }