+++ packages/language-php/grammars/tree-sitter/queries/highlights.scm
@@ -408,7 +408,11 @@
 (const_declaration (const_element) @variable.other.constant.php)
 
 ((name) @constant.language.php
- (#match? @constant.language.php "^__[A-Z][A-Z\d_]+__$"))
+  (#match? @constant.language.php "\\b(__(FILE|DIR|FUNCTION|CLASS|METHOD|LINE|NAMESPACE)__|ON|OFF|YES|NO|NL|BR|TAB)\\b")
+  (#set! capture.final))
+
+((name) @constant.language.php
+  (#match? @constant.language.php "^__[A-Z][A-Z\d_]+__$"))
 
 (argument
   name: (_) @variable.other.named-argument.php
@@ -419,12 +423,12 @@
 
 (string
   "'" @punctuation.definition.string.begin.php
-  (string_value)?
+  (string_content)?
   "'" @punctuation.definition.string.end.php) @string.quoted.single.php
 
 (encapsed_string
   "\"" @punctuation.definition.string.begin.php
-  (string_value)?
+  (string_content)?
   "\"" @punctuation.definition.string.end.php) @string.quoted.double.php
 
 (encapsed_string
@@ -489,7 +493,33 @@
 ; =========
 
 (boolean) @constant.language.boolean._TEXT_.php
+
+; In PHP, `true` is technically no different from `TRUE` or `True`; they appear
+; to be implemented as case-insensitive language constants. `tree-sitter-php`
+; treats `true` as a `boolean` node, but `TRUE` and `True` as `name` nodes.
+;
+; This is silly — but, as usual, the origin of the silliness is PHP.
+;
+; `#match?` is case-sensitive and there's no way to opt out of that. At some
+; point in the future, we might be able to fix that by defining a custom
+; predicate for case-insensitive `#match?`; for now we'll just write a silly
+; regex.
+
+((name) @constant.language.boolean.true.php
+  (#match? @constant.language.boolean.true.php "^[Tt][Rr][Uu][Ee]$")
+  (#set! capture.final))
+
+((name) @constant.language.boolean.false.php
+  (#match? @constant.language.boolean.false.php "^[Ff][Aa][Ll][Ss][Ee]$")
+  (#set! capture.final))
+
 (null) @constant.language.null.php
+
+; Likewise, `null` is a case-insensitive constant. `NULL` and `null` are common;
+; others are less common, but we might as well cover them.
+((name) @constant.language.null.php
+  (#match? @constant.language.null.php "^[Nn][Uu][Ll][Ll]")
+  (#set! capture.final))
 
 (integer) @constant.numeric.decimal.integer.php
 (float) @constant.numeric.decimal.float.php
@@ -558,7 +588,7 @@
 (conditional_expression
   ["?" ":"] @keyword.operator.ternary.php)
 
-(unary_op_expression "@" @keyword.operator.error-control.php)
+(error_suppression_expression "@" @keyword.operator.error-control.php)
 
 [
   "=="