+++ packages/language-typescript/grammars/common/tags.scm
@@ -2,16 +2,16 @@
   name: (identifier) @name
   (#set! symbol.tag "function")) @definition.function
 
-(function
+(function_expression
   name: (identifier) @name
   (#set! symbol.tag "function")) @definition.function
 
 (method_definition
-  name: (property_identifier) @name
+  name: [(property_identifier) (private_property_identifier)] @name
   (#set! symbol.tag "method")) @definition.method
 
 (abstract_method_signature
-  name: (property_identifier) @name
+  name: [(property_identifier) (private_property_identifier)] @name
   (#set! symbol.tag "method")) @definition.method
 
 (class_declaration
@@ -36,7 +36,7 @@
   (lexical_declaration
     (variable_declarator
       name: (identifier) @name
-      value: [(arrow_function) (function)]) @definition.function)
+      value: [(arrow_function) (function_expression)]) @definition.function)
   (#strip! @doc "^[\\s\\*/]+|^[\\s\\*/]$")
   (#select-adjacent! @doc @definition.function)
 )